#13 - 테스트 코드 네이밍 변경
5c407348 commit에서 변경한 네이밍과 변경된 사항에 맞춰 테스트 코드 변경
